[ELF] Change type of `OutputSection::setType()` argument to int64_t
Type of `OutputSection::_type` field is int64_t. This change makes
the field's and the argument's types consistent and allows to assign
full range of values to the `OutputSection::_type` field.

llvm-svn: 233617

+++ lld/lib/ReaderWriter/ELF/SectionChunks.h
@@ -480,7 +480,7 @@ public:
 
   void setFlag(uint64_t flags) { _flags = flags; }
 
-  void setType(int16_t type) { _type = type; }
+  void setType(int64_t type) { _type = type; }
 
   range<ChunkIter> sections() { return _sections; }
